+++ frontend/javascripts/oxalis/view/components/command_palette.tsx
